+++ packages/hoprd/src/index.ts
@@ -138,9 +138,8 @@ const argv = yargsInstance
   })
   .option('apiToken', {
     string: true,
-    describe: 'A REST API token and admin panel password for user authentication [env: HOPRD_API_TOKEN]',
-    default: undefined,
-    conflicts: 'testNoAuthentication'
+    describe: 'A REST API token and for user authentication [env: HOPRD_API_TOKEN]',
+    default: undefined
   })
   .option('healthCheck', {
     boolean: true,
@@ -203,6 +202,11 @@ const argv = yargsInstance
       'Allow connections to other nodes running on private addresses [env: HOPRD_ALLOW_PRIVATE_NODE_CONNECTIONS]',
     default: false
   })
+  .option('disableApiAuthentication', {
+    boolean: true,
+    describe: 'Disable authentication for the API endpoints [env: HOPRD_DISABLE_API_AUTHENTICATION]',
+    default: false
+  })
   .option('testAnnounceLocalAddresses', {
     hidden: true,
     boolean: true,
@@ -221,12 +225,6 @@ const argv = yargsInstance
     describe: 'weaker crypto for faster node startup [env: HOPRD_TEST_USE_WEAK_CRYPTO]',
     default: false
   })
-  .option('testNoAuthentication', {
-    hidden: true,
-    boolean: true,
-    describe: 'no remote authentication for easier testing [env: HOPRD_TEST_NO_AUTHENTICATION]',
-    default: undefined
-  })
   .option('testNoDirectConnections', {
     hidden: true,
     boolean: true,
@@ -239,13 +237,6 @@ const argv = yargsInstance
     boolean: true,
     describe:
       'NAT traversal testing: prevent nodes from establishing direct TCP connections [env: HOPRD_TEST_NO_WEB_RTC_UPGRADE]',
-    default: false
-  })
-  .option('testNoUPNP', {
-    hidden: true,
-    boolean: true,
-    describe:
-      'NAT traversal testing: disable automatic detection of external IP address using UPNP [env: HOPRD_TEST_NO_UPNP]',
     default: false
   })
   .option('heartbeatInterval', {
@@ -315,8 +306,7 @@ function generateNodeOptions(environment: ResolvedEnvironment): HoprOptions {
       announceLocalAddresses: argv.testAnnounceLocalAddresses,
       preferLocalAddresses: argv.testPreferLocalAddresses,
       noWebRTCUpgrade: argv.testNoWebRTCUpgrade,
-      noDirectConnections: argv.testNoDirectConnections,
-      noUPNP: argv.testNoUPNP
+      noDirectConnections: argv.testNoDirectConnections
     }
   }
 
@@ -402,7 +392,7 @@ async function main() {
     }
   }
 
-  if (!argv.testNoAuthentication && argv.api) {
+  if (!argv.disableApiAuthentication && argv.api) {
     if (argv.apiToken == null) {
       throw Error(`Must provide --apiToken when --api is specified`)
     }
@@ -411,7 +401,7 @@ async function main() {
     }
   }
 
-  const apiToken = argv.testNoAuthentication ? null : argv.apiToken
+  const apiToken = argv.disableApiAuthentication ? null : argv.apiToken
 
   // We need to setup the admin server before the HOPR node
   // as if the HOPR node fails, we need to put an error message up.