+++ scripts/run-integration-tests-source.sh
@@ -184,7 +184,7 @@ function setup_node() {
   if [[ -n "${api_token}" ]]; then
     additional_args="--api-token='${api_token}' ${additional_args}"
   else
-    additional_args="--testNoAuthentication ${additional_args}"
+    additional_args="--disableApiAuthentication ${additional_args}"
   fi
 
   # Remove previous logs to make sure the regex does not match
@@ -220,7 +220,6 @@ function setup_node() {
       --testAnnounceLocalAddresses \
       --testPreferLocalAddresses \
       --testUseWeakCrypto \
-      --testNoUPNP \
       --allowLocalNodeConnections \
       --allowPrivateNodeConnections \
       ${additional_args} \